requirements.txt
altgraph==0.17.4
amqp==5.3.1
APScheduler==3.11.0
argh==0.31.3
asgiref==3.9.0
astroid==3.3.10
attrs==26.1.0
autopep8==2.3.1
babel==2.17.0
bcrypt==4.3.0
billiard==4.2.1
black==24.8.0
boto3==1.35.24
botocore==1.35.99
Brotli==1.1.0
celery==5.4.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
code2flow==2.5.1
contourpy==1.3.2
coverage==7.9.2
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.8.0rc2
diff-match-patch==20241021
dill==0.4.0
dj-database-url==3.1.2
Django==4.2.11
django-allauth==64.1.0
django-appconf==1.1.0
django-autoslug==1.9.8
django-axes==6.5.1
django-bootstrap-modal-forms==3.0.5
django-celery-email==3.0.0
django-cleanup==8.0.0
django-cors-headers==4.4.0
django-countries==7.6.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-import-export==4.1.1
django-model-utils==5.0.0
django-phonenumber-field==8.0.0
django-ratelimit==4.0.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-storages==1.14.4
django-tables2==2.7.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.29.0
drf-yasg==1.21.7
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==30.0.0
flake8==7.1.1
flower==2.0.1
fonttools==4.58.2
graphviz==0.21
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
kombu==5.5.4
matplotlib==3.10.3
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.0
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
phonenumbers==8.13.47
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
pyflakes==3.2.0
Pygments==2.19.2
pygraphviz==1.14
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.9.0
pylint==3.3.7
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.4.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.37.0
reportlab==4.4.4
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.30.0
s3transfer==0.10.4
sentry-sdk==2.13.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
tablib==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
tomlkit==0.13.3
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
watchdog==5.0.2
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.7.0
zopfli==0.2.3.post1
